supervisor/requirements.txt
fastapi
uvicorn
a2a-sdk
ag-ui-protocol
langgraph
langchain
langchain_openai
langchain-community
httpx
python-dotenv